# Ceramic_Detection/Ceramic_Detection/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ceramic Detection API",
        "description": "API documentation for the Ceramic Detection application",
        "version": "1.0.0"
    },
    "paths": {
...
        "/gallery": {
            "get": {
                "tags": [
                    "Ceramic"
                ],
                "summary": "Get ceramic gallery",
                "description": "Retrieve a paginated list of ceramics with optional category and origin filters",
                "operationId": "fba6ddb948ce140fa47cf85cca945142",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "description": "Filter by ceramic category",
                            "type": "string"
                        }
                    },
                    {
                        "name": "origin",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "description": "Filter by ceramic origin",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ceramics": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "categories": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "origins": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/dashboard": {
            "get": {
                "tags": [
                    "Ceramic"
                ],
                "summary": "Get user dashboard",
                "description": "Retrieve classification history for the authenticated user",
                "operationId": "2beff246d8d78e60a41b9f75d4619355",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "classifications": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}